split out ui related stuff to ease codebase navigation
